plat/native/arm64: Implement ectx
Migrate the arm64 extended context from arch/arm and plat/common
to plat/native as uk_plat_native_ectx. Add the PAL headers that
wrap around the implementation of plat/native.

Checkpatch-Ignore: LONG_LINE
Checkpatch-Ignore: OPEN_ENDED_LINE
