+++ arch/arm/Makefile.uk
@@ -11,7 +11,6 @@ $(eval $(call addlib,libcontext))
 
 LIBCONTEXT_CINCLUDES-y += -I$(UK_PLAT_COMMON_BASE)/include
 LIBCONTEXT_SRCS-y += $(LIBCONTEXT_BASE)/ctx.c|isr
-LIBCONTEXT_SRCS-y += $(LIBCONTEXT_BASE)/ectx.c|isr
 LIBCONTEXT_SRCS-$(CONFIG_ARCH_ARM_64) += $(LIBCONTEXT_BASE)/arm64/execenv.S|arm64
 LIBCONTEXT_SRCS-$(CONFIG_ARCH_ARM_64) += $(LIBCONTEXT_BASE)/arm64/ctx.S|arm64
 LIBCONTEXT_SRCS-$(CONFIG_ARCH_ARM_64) += $(LIBCONTEXT_BASE)/arm64/tls.c|arm64
