+++ arch/arm/arm64/include/uk/asm/ctx.h
@@ -35,10 +35,6 @@
 #error Do not include this header directly
 #endif
 
-/* 32 * 128-bit NEON registers + 4 bytes (FPSR) + 4 bytes (FPCR) */
-#define UKARCH_ECTX_SIZE	520
-#define UKARCH_ECTX_ALIGN	16
-
 /* Stack needs to be aligned to 16 bytes */
 #define UKARCH_SP_ALIGN		16
 #define UKARCH_SP_ALIGN_MASK	(UKARCH_SP_ALIGN - 1)
