+++ plat/native/Makefile.uk
@@ -21,6 +21,7 @@ CXXINCLUDES-$(CONFIG_LIBUKPLAT_NATIVE_PAL) += -I$(LIBUKPLAT_NATIVE_BASE)/pal/arc
 LIBUKPLAT_NATIVE_CINCLUDES-y += -I$(UK_PLAT_COMMON_BASE)/include
 
 ifeq ($(CONFIG_ARCH_ARM_64), y)
+LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_ECTX)   += $(LIBUKPLAT_NATIVE_BASE)/arch/arm64/ectx.c|isr
 LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_PAGING) += $(LIBUKPLAT_NATIVE_BASE)/arch/arm64/paging.c|isr
 LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_PAGING) += $(LIBUKPLAT_NATIVE_BASE)/arch/arm64/pt.c|isr
 LIBUKPLAT_NATIVE_SRCS-$(CONFIG_LIBUKPLAT_NATIVE_SYSCTX) += $(LIBUKPLAT_NATIVE_BASE)/arch/arm64/sysctx.c|isr
